requirements.txt
openai
langchain-openai
flask
python-dotenv
SQLAlchemy
flask_sqlalchemy
bcrypt
flask-mail